Remove usage of experimental typeNameOf function
typeNameOf from KotlinPoet is annotated with ExperimentalStdlibApi, so
requires all callers also opt-in or also be annotated with
ExperimentalStdlibApi. Usage of this one helper was causing a
proliferation of classes and methods to be annotated, when most of the
actual usage of typeNameOf was not really much more concise; other than
the tests, there were no instances of nullable or parameterized types,
so switching to the asTypeName extension method on KClass is not too
verbose.