Smoke Test that uses CodeGen in the context of a SB App.
Context
=======

This commit adds a smoke test that will use the plugin in the context of a Spring Boot Application.
It also adds some DGS specific features, such as the `com.netflix.graphql.dgs:graphql-dgs-extended-scalars`.
The test will apply against a vanilla Java project as well as a Kotlin
project.

Rationale
=========

The tests automatically executed by the build of this plugin will not necessarily considered the common/expected
usage pattern in the context of the DGS Framework. That is, in the context of a Spring Boot Application
that leverages other components of the DGS Framework.

Scope of this Change
====================

The scope of this change is limited to the addition of the `CodegenGradlePluginSpringBootSmokeTest`, refactoring
`CodegenGradlePluginTest` to Kotlin, and refactoring the build such that
we remove the unused `functionalTest` facet. The latter was causing
additional problems with the default behaviour of the Gradle Test Kit.

The CodegenGradlePluginSpringBootSmokeTest
==========================================

The Spring Boot Application used by the test if fairly simple, as well
as the test that it internally will execute as part of its validation.
In its current form it only asserts that the Spring Boot application is
brought up correctly, which includes loading the default
Auto-Configurations. In the future this could be improved, that said,
the scope of such test is not necessarily to assert all permutations of
the settings for the Gradle Plugin.