Test Example Application as part of the CI Process
Summary
=======

This commit, via the ci-beta.yml workflow action, will use the build's
Codegen snapshot version and apply it to the example applications as
defined in the scripts/config.yml file, included in this PR.
It will then execute a full build of each of them as part of the verification process.

Implementation
==============

Via the `scripts/test-examples.py` we are adding the ability to download
defined git repositories that we want to use to assert the behaviour of the Codegen plugin changes.
The same script will update the downloaded projects' files,
in this case the `settings.gradle.kts` file, to override
the artifact such projects use. This enables us to force the build to
use the locally published Codegen snapshot.

Requirements of Example Projects.

1. Should build with the same JDK as the Codegen project. In this case JDK 1.8.
2. Should not be a multi-module project since we are overriding the
   `settings.gradle.kts` file.

Running The Example Tests Locally
=================================

1. You will need Python3 and have `pip3` and `python3` available in your
   path.
2. You will need to install the Python3 required modules defined in
   `scripts/requirements.txt`. You can just run `make install-py-libs`.
3. Run the `make test-examples` tasks.

You can also run `./scripts/test-examples.py` manually. Explore the
options with the `-h` flag.