Fix to include interface methods on subprojections when it's not a fragment
Currently `ClientApiGenerator.createSubProjectionType` ignores the interface fields when the type implements a interface. For example, when passing the following schema, the `name` method for the `Search_DirectorProjection` subprojecion class is not being generated.

```graphql
type Query {
    search(title: String): [Show]
}

interface Show {
    title: String
    director: Director
}

interface Person {
    name: String
}

type Director implements Human {
    name: String
    shows: [Show]
}
```

This commit fixes that by checking if there's a fragment in the parent class referring to the type of the subprojection type being generated. If so, it filters out the interface fields. Otherwise, the interface fields aren't excluded.