Adding a graphql-dgs-codegen-client-core module
What
====

This commit adds a new module, the `graphql-dgs-codegen-client-core`, that offers classes, and other artifacts, that complement the code generated by _DGS Codegen_.
The plugin, by default, will add the artifact automatically to the _implementation_ scope.

As a Developer
=============

As a Developer you will be able to leverage the `graphql-dgs-codegen-client-core`.
Optionally you have the option to disable  the inclusion of such dependency by setting
the `codegen.clientCoreConventionsEnabled` flag to `false` as follows.

```
codegen.clientCoreConventionsEnabled = true
```

or

```
codegen {
    clientCoreConventionsEnabled = true
}
```