Generate nested input types when required
Sometimes users want to disable the generation of all _data types_, via the `generateDataTypes` flag.
This is convenient when the schema expresses a large number of such types but the user is only in a subsection.
When generating such subsection we need to include the nested input and enumerations that are part of the sub-graph
the user is interested in.