Fields with basic types should support projections
The _Client API codegen_ will not generate projection types for _basic type fields_,
such as Strings, Booleans, Integers, or Floats, or arrays of any of these types.

Let's say we have the following schema:

```
type Query {
    someField: Foo
}

type Foo {
    stringField(arg: Boolean): String
}
```

According to the schema we should expect a _stringField projection_ on the `someField` field.
But currently this is not the case.

When generating the root projections, as well as sub-projections, we
need to consider the _fields_ with basic types that present input arguments.

To be more specific, and leveraging the schema mentioned above, the `SomeFieldProjectionRoot` _projection type_
should include a `stringField(Boolean arg)` method definition.

The changes on this PR will generate the following `stringField(Boolean arg)` method...

```
public SomeField_StringFieldProjection stringField(Boolean arg) {
    SomeField_StringFieldProjection projection = new SomeField_StringFieldProjection(this, this);
    getFields().put("stringField", projection);
    getInputArguments().computeIfAbsent("stringField", k -> new ArrayList<>());
    InputArgument argArg = new InputArgument("arg", arg);
    getInputArguments().get("stringField").add(argArg);
    return projection;
}
```

In addition to the `stringFeild()` method that will already exist in the `SomeFieldProjectionRoot`, such as the one
below.

```
public SomeFieldProjectionRoot stringField() {
    getFields().put("stringField", null);
    return this;
}
```

Note that the empty argument method will still return the _root projection_ class, in this
case `SomeFieldProjectionRoot` while the new field, the one with the argument, will return a different type,
in this case `SomeField_StringFieldProjection`.
This is the same behavior for complex types and although arguably both
could return the new type, `SomeField_StringFieldProjection`, this will
be backwards incompatible.