Fields with scalars should support projections
Current Behavior
================

The _Client API codegen_ will not generate projection types for _scalars_. Let's say we have the following schema:

```
type Query {
    ping: Foo
}

type Foo {
    someField(arg: Boolean): Long
}

scalar Long
```

According to the schema we should expect a _someField projection_, but currently this is not the case.

Suggested Improvement
=====================

When generating the root projections, as well as sub-projections, we
need to consider the _fields_ with basic types that present input arguments.

To be more specific, and leveraging the schema mentioned above, the `SomeFieldProjectionRoot` _projection type_
should include a `ping(Boolean arg)` method definition.

The changes on this PR will generate the following `ping(Boolean arg)` method...

```
public SomeField_PingProjection ping(Boolean arg) {
        SomeField_PingProjection projection = new SomeField_PingProjection(this, this);
        getFields().put("ping", projection);
        getInputArguments().computeIfAbsent("ping", k -> new ArrayList<>());
        InputArgument argArg = new InputArgument("arg", arg);
        getInputArguments().get("ping").add(argArg);
        return projection;
}
```

In addition to the `ping()` method that will already exist in the `SomeFieldProjectionRoot`, such as the one
below.

```
public SomeField_PingProjection ping() {
    SomeField_PingProjection projection = new SomeField_PingProjection(this, this);
    getFields().put("ping", projection);
    return projection;
}
```