Support Projections on complex Entity fields
This commit will resolve projection on _Entity_ fields with complex
types. For example, let's say we have a `Foo` entity with a field
projection that returns a `Bar`.

```
type Foo @key(fields:"id") {
    id: ID
    bar(arg: [String!]): Bar
}

type Bar {
    name: String
}
```

In the past the root projection for the entity will only generate a
`bar` method with no arguments. With this change it will have, in
addition, the `bar` method that supports the arguments for the
projection.