Upgrade Kotlin, Kotlin Poet, and Kotlinter
Upgrading to Kotlin 1.7.0
=========================

Upgrade to 'com.squareup:kotlinpoet:1.12.0' and 'com.github.ajalt.clikt:clikt:3.5.0' brought Kotlin 1.7.0 so decided
to upgrade the project as well. If we want to revert moving to Kotlin
1.7.0 we will need to pin clickt and kotlinpoet to 3.4.+ and 1.11.+ respectively.

Upgrading to Kotlinter 3.11.1
=============================

Moving to kotlinter 3.11.1 since 3.6.0 was creating problems with tests
since, it appeared, that it was taking into account string template blocks
as part of the code.

Other Fixes
===========

* Make sure we run the CI job on every push and not only when the PR is opened
* Remove the `.env` file, use `sdk env` instead