Fix Kotlin data classes when generateAllConstructors is disabled
Fix Kotlin data classes when the `generateAllConstructors` is disabled.
The path I took was to generate _mutable_ Kotlin Data Classes since then
the constructor with all the options is not needed and the type
properties can be set after the constructor is executed.

We should reconsider if we have the Kotlin generator consider `generateAllConstructors` at all.
I think it would be a better option if we have Java and Kotlin specific
flags. In this case it could be, for Java, `generateAllConstructors`;
while for Kotlin we could have `mutableDataTypes`.

Upgrading to Kotlin 1.7.0
=========================

Upgrade to 'com.squareup:kotlinpoet:1.12.0' and 'com.github.ajalt.clikt:clikt:3.5.0' brought Kotlin 1.7.0 so decided
to upgrade the project as well. If we want to revert moving to Kotlin
1.7.0 we will need to pin clickt and kotlinpoet to 3.4.+ and 1.11.+ respectively.

Upgrading to Kotlinter 3.11.1
=============================

Moving to kotlinter 3.11.1 since 3.6.0 was creating problems with tests
since, it appeared, that it was taking into account string template blocks
as part of the code.

Other Fixes
===========

* Make sure we run the CI job on every push and not only when the PR is opened
* Remove the `.env` file, use `sdk env` instead