The generateAllConstructor flag will only affect Java Types
The case of supporting `generateAllConstructor` in Kotlin is very week
since GraphQL Types are generated as _Kotlin Data Types. Because of that
we decided to not support it as part of Kotlin. We also decided to
rename the flag from `generateAllConstructor` to `javaGenerateAllConstructor`,
such that it describes it is explicit to Java.