Align the dgs-codegen-core and dgs-codegen-shared-core
Objective
=========

This commit aligns the dgs-codegen-core and dgs-codegen-shared-core
such that they share the same graphql-java version.
This will prevent deviations that might cause confusion.

Changes
=======

This commit adds the platform bom as an `api` and `implementation`
platform for all sub-modules, simplifying the coordination of the
DGS BOM version that is used. In addition, the `graphql-dgs-codegen-core`
will now fetch transitively the `graphql-java` dependency from
`graphql-dgs-codegen-shared-core` ensuring both use the same version.

Other
=====

* Updating the Dependency Locks